Moved ShareFetchMetadata to common/requests
