+++ clients/src/main/java/org/apache/kafka/clients/consumer/internals/ShareSessionHandler.java
@@ -22,6 +22,7 @@ import org.apache.kafka.common.TopicPartition;
 import org.apache.kafka.common.Uuid;
 import org.apache.kafka.common.protocol.Errors;
 import org.apache.kafka.common.requests.ShareAcknowledgeResponse;
+import org.apache.kafka.common.requests.ShareFetchMetadata;
 import org.apache.kafka.common.requests.ShareFetchResponse;
 import org.apache.kafka.common.utils.LogContext;
 import org.apache.kafka.common.utils.Utils;
@@ -81,96 +82,6 @@ public class ShareSessionHandler {
         this.memberId = memberId;
         this.nextMetadata = ShareFetchMetadata.initialEpoch(memberId);
         this.sessionPartitions = new LinkedHashMap<>();
-    }
-
-    public static class ShareFetchMetadata {
-        /**
-         * The first epoch. When used in a ShareFetch request, indicates that the client
-         * wants to create a session.
-         */
-        public static final int INITIAL_EPOCH = 0;
-
-        /**
-         * An invalid epoch. When used in a ShareFetch request, indicates that the client
-         * wants to close an existing session.
-         */
-        public static final int FINAL_EPOCH = -1;
-
-        /**
-         *
-         */
-        public boolean isNewSession() {
-            return epoch == INITIAL_EPOCH;
-        }
-
-        /**
-         * Returns the next epoch.
-         *
-         * @param prevEpoch The previous epoch.
-         * @return          The next epoch.
-         */
-        public static int nextEpoch(int prevEpoch) {
-            if (prevEpoch < 0) {
-                // The next epoch after FINAL_EPOCH is always FINAL_EPOCH itself.
-                return FINAL_EPOCH;
-            } else if (prevEpoch == Integer.MAX_VALUE) {
-                return 1;
-            } else {
-                return prevEpoch + 1;
-            }
-        }
-
-        /**
-         * The member ID.
-         */
-        private final Uuid memberId;
-
-        /**
-         * The share session epoch.
-         */
-        private final int epoch;
-
-        public ShareFetchMetadata(Uuid memberId, int epoch) {
-            this.memberId = memberId;
-            this.epoch = epoch;
-        }
-
-        public static ShareFetchMetadata initialEpoch(Uuid memberId) {
-            return new ShareFetchMetadata(memberId, INITIAL_EPOCH);
-        }
-
-        public ShareFetchMetadata nextEpoch() {
-            return new ShareFetchMetadata(memberId, nextEpoch(epoch));
-        }
-
-        public ShareFetchMetadata nextCloseExistingAttemptNew() {
-            return new ShareFetchMetadata(memberId, INITIAL_EPOCH);
-        }
-
-        public ShareFetchMetadata finalEpoch() {
-            return new ShareFetchMetadata(memberId, FINAL_EPOCH);
-        }
-
-        public Uuid memberId() {
-            return memberId;
-        }
-
-        public int epoch() {
-            return epoch;
-        }
-
-        public String toString() {
-            StringBuilder bld = new StringBuilder();
-            bld.append("(memberId=").append(memberId).append(", ");
-            if (epoch == INITIAL_EPOCH) {
-                bld.append("epoch=INITIAL)");
-            } else if (epoch == FINAL_EPOCH) {
-                bld.append("epoch=FINAL)");
-            } else {
-                bld.append("epoch=").append(epoch).append(")");
-            }
-            return bld.toString();
-        }
     }
 
     public static class ShareFetchRequestData {
